# shared/openapi.yaml
openapi: 3.1.0
...
info:
  title: Syl API
  version: 1.0.0
  summary: The contract between Syl's service, her iOS app, and the web admin.
  description: |
    # Conventions

    ## Envelope

    Every response body — success or failure — is one of:

    ```json
    { "success": true,  "data": { ... } }
    { "success": false, "error": { "code": "NOT_FOUND", "message": "...", "retryable": false } }
    ```

    There is no third shape. A client that cannot parse one of these two has
    hit something that is not Syl (a proxy, a captive portal, a Tailscale
    error page) and should treat it as a transport failure, not an API error.

    ## Field naming

    **camelCase for every JSON field, with exactly one exception: `ttl_ms` on
    the `presence` frame.** That field is quoted as literal JSON in the
    character proposal and in `syl-001.2.2`, three separate sources pin the
    spelling, and silently "fixing" it is how a contract stops matching the
    thing it was derived from. It is the only snake_case identifier on the
    wire. Do not add a second one.

    Note that the service's SQLite columns are snake_case (child B). The
    mapping is at the store boundary, not on the wire.

    ## Identifiers

    Every resource id is `syl:<type>:<uuidv7>` — type-prefixed so a dangling
    reference is legible in a log line, UUIDv7 so ids sort by creation time.
    The convention is fixed before the first row is written because every
    graph edge will reference it forever.

...

...

...

...

...

...

...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /goals:
    get:
      tags: [goals]
      operationId: listGoals
      summary: The minimal goal skeleton — S1 scope only.
      description: |
        Goals self-nest through `parentId`; there is no separate
        objective/key-result entity. There is also no percent-complete field:
        self-reported percentages are fiction and they decay. Progress is
        evidenced by what was linked to the goal, and computed.
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
        - name: status
          in: query
          required: false
          schema: { $ref: "#/components/schemas/GoalStatus" }
      responses:
        "200":
          description: A page of goals.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/OkEnvelope"
                  - type: object
                    required: [data]
                    properties:
                      data: { $ref: "#/components/schemas/GoalPage" }
        default: { $ref: "#/components/responses/Error" }
    post:
      tags: [goals]
      operationId: createGoal
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateGoalRequest" }
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/OkEnvelope"
                  - type: object
                    required: [data]
                    properties:
                      data: { $ref: "#/components/schemas/Goal" }
        default: { $ref: "#/components/responses/Error" }
  /goals/{goalId}:
    get:
      tags: [goals]
      operationId: getGoal
      parameters:
        - $ref: "#/components/parameters/GoalId"
      responses:
        "200":
          description: The goal.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/OkEnvelope"
                  - type: object
                    required: [data]
                    properties:
                      data: { $ref: "#/components/schemas/Goal" }
        default: { $ref: "#/components/responses/Error" }

  # ------------------------------------------------------------- devices ---
    patch:
      tags: [goals]
      operationId: updateGoal
      summary: Reword a goal, or move it between states.
      description: |
        The one write a goal had no way to express. `status` carries achieving
        it and abandoning it as well as reactivating a dormant one, because
        they are the same operation on the same row and separate endpoints
        would have to agree about what each leaves behind.
      parameters:
        - $ref: "#/components/parameters/GoalId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateGoalRequest" }
      responses:
        "200":
          description: Updated.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/OkEnvelope"
                  - type: object
                    required: [data]
                    properties:
                      data: { $ref: "#/components/schemas/Goal" }
        default: { $ref: "#/components/responses/Error" }
...
  /jobs:
    get:
      tags: [jobs]
      operationId: listJobs
      summary: What is scheduled. Backs the admin's "Next" view.
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
        - name: state
          in: query
          required: false
          schema: { $ref: "#/components/schemas/JobState" }
        - name: kind
          in: query
          required: false
          schema: { $ref: "#/components/schemas/JobKind" }
      responses:
        "200":
          description: A page of jobs.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/OkEnvelope"
                  - type: object
                    required: [data]
                    properties:
                      data: { $ref: "#/components/schemas/JobPage" }
        default: { $ref: "#/components/responses/Error" }
  /jobs/{jobId}:
    get:
      tags: [jobs]
      operationId: getJob
      parameters:
        - $ref: "#/components/parameters/JobId"
      responses:
        "200":
          description: The job.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/OkEnvelope"
                  - type: object
                    required: [data]
                    properties:
                      data: { $ref: "#/components/schemas/Job" }
        default: { $ref: "#/components/responses/Error" }
  /jobs/{jobId}/runs:
    get:
      tags: [jobs]
      operationId: listJobRuns
      summary: Backs the admin's "Ran" and "Failed" views.
      parameters:
        - $ref: "#/components/parameters/JobId"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: A page of runs, newest first.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/OkEnvelope"
                  - type: object
                    required: [data]
                    properties:
                      data: { $ref: "#/components/schemas/RunPage" }
        default: { $ref: "#/components/responses/Error" }
  /runs/{runId}:
    get:
      tags: [jobs]
      operationId: getRun
      summary: One run with its ordered steps.
      parameters:
        - $ref: "#/components/parameters/RunId"
      responses:
        "200":
          description: The run.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/OkEnvelope"
                  - type: object
                    required: [data]
                    properties:
                      data: { $ref: "#/components/schemas/Run" }
        default: { $ref: "#/components/responses/Error" }

  # ---------------------------------------------------------------- logs ---
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        Client-generated, opaque, unique per logical operation. Required on
        every write — the mobile outbox retries by design and will duplicate
        without it.
      schema: { type: string, minLength: 8, maxLength: 128 }
    Cursor:
      name: cursor
      in: query
      required: false
      description: Opaque page cursor from the previous page's `nextCursor`.
      schema: { type: string }
    Limit:
      name: limit
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
...
    GoalId:
      name: goalId
      in: path
      required: true
      schema: { $ref: "#/components/schemas/Id" }
...
    JobId:
      name: jobId
      in: path
      required: true
      schema: { $ref: "#/components/schemas/Id" }
    RunId:
      name: runId
      in: path
      required: true
      schema: { $ref: "#/components/schemas/Id" }
  responses:
    Error:
      description: |
        Any failure. The HTTP status is advisory; `error.code` is the
        contract. `error.retryable` tells the client's backoff whether to try
        again — the Tailscale extension is torn down when idle, so the first
        request after a wake can fail while the tunnel re-establishes, and
        treating that as "server down" makes a healthy app feel broken.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
  schemas:
...
    Id:
      type: string
      description: "`syl:<type>:<uuidv7>` — type-prefixed, time-sortable."
      pattern: "^syl:[a-z][a-z_]*:[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$"
      examples: ["syl:reminder:0198f2c1-4a3b-7d21-9f00-1a2b3c4d5e6f"]
    Instant:
      type: string
      format: date-time
      description: RFC 3339, UTC, millisecond precision. Never a fixed offset.
      examples: ["2026-08-09T07:00:03.114Z"]
    WallTime:
      type: string
      description: |
        24-hour local wall-clock time. Stored alongside an IANA `tz`; the
        instant is a materialisation of the next occurrence, never the source
        of truth. Storing only the instant is the fixed-offset bug in a
        different costume.
      pattern: "^([01][0-9]|2[0-3]):[0-5][0-9]$"
      examples: ["07:00"]
    Timezone:
      type: string
      description: IANA zone name. Never a fixed UTC offset.
      examples: ["America/Chicago"]
    OkEnvelope:
      type: object
      required: [success]
      properties:
        success: { const: true }
    ErrorEnvelope:
      type: object
      required: [success, error]
      properties:
        success: { const: false }
        error: { $ref: "#/components/schemas/ApiError" }
    ErrorCode:
      type: string
      description: |
        The typed code is the contract; the HTTP status is advisory. Clients
        branch on this, never on the status line.

        `PAIRING_CODE_EXPIRED` and `PAIRING_CODE_ALREADY_USED` are the only
        places the service tells one authentication failure from another, and
        they are safe because **both are reachable only by presenting a code
        that matches a stored one**. A wrong guess, and an attempt made while
        no code is live, are the same `UNAUTHORIZED`. So the extra detail
        reaches the person holding the code and nobody else — which is what
        lets a pairing screen on a phone say something more useful than "that
        did not work".
      enum:
        - VALIDATION_FAILED
        - UNAUTHORIZED
        - PAIRING_CODE_EXPIRED
        - PAIRING_CODE_ALREADY_USED
        - FORBIDDEN
        - NOT_FOUND
        - CONFLICT
        - IDEMPOTENCY_KEY_REUSE
        - IDEMPOTENCY_KEY_REQUIRED
        - DEFERRAL_NOT_LATER
        - RRULE_UNSUPPORTED
        - UNKNOWN_JOB_KIND
        - DEVICE_TOKEN_INVALID
        - QUIET_HOURS
        - RATE_LIMITED
        - UPSTREAM_UNAVAILABLE
        - INTERNAL
    ApiError:
      type: object
      required: [code, message, retryable]
      properties:
        code: { $ref: "#/components/schemas/ErrorCode" }
        message:
          type: string
          description: Human-readable, for a log line or an admin screen. Never parsed.
        retryable:
          type: boolean
          description: |
            Whether the client's backoff should try again. `RATE_LIMITED`,
            `UPSTREAM_UNAVAILABLE` and `INTERNAL` are retryable; a validation
            failure never is. Auth and billing failures halt rather than
            retry — retrying an auth failure fifty times is the worst
            possible response to a key having shadowed the subscription login.
        details:
          type: [object, "null"]
          description: Structured context. `VALIDATION_FAILED` carries `{ field, reason }[]`.
          additionalProperties: true
        retryAfterMs:
          type: [integer, "null"]
          description: Server-suggested backoff floor, when it knows one.
...
    GoalStatus:
      type: string
      description: |
        `abandoned` is a first-class, non-shameful outcome. `dormant` is a
        real state and reactivating a dormant goal restores its history
        intact. The three answers Syl offers — still mine / not now / done
        with it — map to `active` / `dormant` / `abandoned`.
      enum: [proposed, active, dormant, achieved, abandoned]
    Goal:
      type: object
      required:
        [id, parentId, title, why, targetDate, metricKey, targetValue, cadenceDays, status, statusReason, createdAt, updatedAt]
      properties:
        id: { $ref: "#/components/schemas/Id" }
        parentId:
          oneOf:
            - { $ref: "#/components/schemas/Id" }
            - { type: "null" }
          description: Goals self-nest. There is no separate objective entity.
        title: { type: string }
        why:
          type: [string, "null"]
          description: The only optional field Syl should ever push for.
        targetDate:
          type: [string, "null"]
          description: "`YYYY-MM-DD`. The horizon (life / year / season / month) is derived from it, not stored."
          pattern: "^[0-9]{4}-[0-9]{2}-[0-9]{2}$"
        metricKey: { type: [string, "null"] }
        targetValue: { type: [number, "null"] }
        cadenceDays:
          type: [integer, "null"]
          description: Drives the silence signal — nothing linked for longer than this is a risk.
        status: { $ref: "#/components/schemas/GoalStatus" }
        statusReason: { type: [string, "null"] }
        createdAt: { $ref: "#/components/schemas/Instant" }
        updatedAt: { $ref: "#/components/schemas/Instant" }
    GoalPage:
      type: object
      required: [items, nextCursor, hasMore]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Goal" }
        nextCursor: { type: [string, "null"] }
        hasMore: { type: boolean }
    CreateGoalRequest:
      type: object
      required: [title]
      properties:
        title: { type: string, minLength: 1 }
        parentId:
          oneOf:
            - { $ref: "#/components/schemas/Id" }
            - { type: "null" }
        why: { type: [string, "null"] }
        targetDate: { type: [string, "null"], pattern: "^[0-9]{4}-[0-9]{2}-[0-9]{2}$" }
        cadenceDays: { type: [integer, "null"], minimum: 1 }
    UpdateGoalRequest:
      # Every field optional, and only the named ones move. A patch that
      # required the whole goal back would silently overwrite whatever the
      # caller did not think to include — the failure where rewording a goal
      # quietly clears its target date.
      type: object
      properties:
        title: { type: string, minLength: 1 }
        why: { type: [string, "null"] }
        targetDate: { type: [string, "null"], pattern: "^[0-9]{4}-[0-9]{2}-[0-9]{2}$" }
        cadenceDays: { type: [integer, "null"], minimum: 1 }
        status: { $ref: "#/components/schemas/GoalStatus" }
        statusReason: { type: [string, "null"] }

    # --------------------------------------------------------- devices ---
...
    JobKind:
      type: string
      description: |
        A closed catalogue. The model may enqueue a job; it may never invent
        a kind. This is the trust boundary pushed down one level — if the
        model could create arbitrary recurring jobs, a prompt injection
        inside an article becomes a job that speaks to him every morning.

        Note the shape of the catalogue: the two kinds with a hard guarantee
        attached, `reminder_delivery` and `maintenance`, are the two that use
        no turns. That is not a coincidence.
      enum:
        - reminder_delivery
        - morning_agenda
        - evening_review
        - heartbeat
        - nightly_consolidation
        - research_brief
        - content_ingestion
        - maintenance
    JobState:
      type: string
      enum: [pending, leased, running, done, failed, abandoned, suspended]
    JobPriority:
      type: string
      description: Background never starts while interactive work is pending.
      enum: [interactive, reminder, scheduled, background]
    JobDeliveryClass:
      type: string
      description: |
        Four values, not three: the proposal's §2 lists `at_least_once` and
        §6 splits it into idempotent and resumable, which behave differently
        after a crash — one is retried from the top, the other resumes from
        its last recorded step. They are different guarantees and get
        different names.
      enum: [at_least_once, at_least_once_resumable, at_most_once, once_per_window]
    CatchUpPolicy:
      type: string
      description: |
        What to do about an instant that passed while we were down.
        `never_expires` fires however late and marks it late — reminders.
        `grace_window` runs if still inside the window, else skips and logs.
        `skip` collapses three missed heartbeats into none.
        `once_per_window` uses a marker row so a second attempt is a no-op.
      enum: [never_expires, grace_window, skip, once_per_window]
    JobCatchUp:
      type: object
      required: [policy]
      properties:
        policy: { $ref: "#/components/schemas/CatchUpPolicy" }
        graceMs: { type: [integer, "null"], minimum: 0 }
        windowStart: { $ref: "#/components/schemas/WallTime" }
        windowEnd: { $ref: "#/components/schemas/WallTime" }
    JobTriggerType:
      type: string
      enum: [wall_clock, interval, event, manual]
    JobTrigger:
      type: object
      required: [type]
      properties:
        type: { $ref: "#/components/schemas/JobTriggerType" }
        wallTime: { $ref: "#/components/schemas/WallTime" }
        tz: { $ref: "#/components/schemas/Timezone" }
        rrule: { type: [string, "null"] }
        intervalMs: { type: [integer, "null"], minimum: 1000 }
        event: { type: [string, "null"] }
    JobBudget:
      type: object
      required: [maxTurns, maxWallClockMs, allowedTools]
      description: |
        `allowedTools` is a property of the job kind, which is what bounds
        what a headless turn can reach. Reminder delivery needs no tools at
        all — and `maxTurns: 0` is the strongest statement in the catalogue,
        because a job that cannot spawn a turn cannot be delayed by a rate
        limit or broken by a model declining to act.
      properties:
        maxTurns: { type: integer, minimum: 0 }
        maxWallClockMs: { type: integer, minimum: 0 }
        allowedTools:
          type: array
          items: { type: string }
    JobLease:
      type: object
      required: [owner, expiresAt]
      description: What makes reboot recovery possible. Recovery runs before scheduling.
      properties:
        owner: { type: string }
        expiresAt: { $ref: "#/components/schemas/Instant" }
    CircuitBreakerState:
      type: string
      enum: [closed, open, half_open]
    CircuitBreaker:
      type: object
      required: [state, consecutiveFailures, openedAt]
      description: N consecutive failures disables the kind and reports once. Nothing retries forever.
      properties:
        state: { $ref: "#/components/schemas/CircuitBreakerState" }
        consecutiveFailures: { type: integer, minimum: 0 }
        openedAt:
          oneOf:
            - { $ref: "#/components/schemas/Instant" }
            - { type: "null" }
    Job:
      type: object
      required:
        [id, kind, state, priority, trigger, deliveryClass, catchUp, budget, lease, circuitBreaker, nextRunAt, lastRunId, speaks, createdAt, updatedAt]
      properties:
        id: { $ref: "#/components/schemas/Id" }
        kind: { $ref: "#/components/schemas/JobKind" }
        state: { $ref: "#/components/schemas/JobState" }
        priority: { $ref: "#/components/schemas/JobPriority" }
        trigger: { $ref: "#/components/schemas/JobTrigger" }
        deliveryClass: { $ref: "#/components/schemas/JobDeliveryClass" }
        catchUp: { $ref: "#/components/schemas/JobCatchUp" }
        budget: { $ref: "#/components/schemas/JobBudget" }
        lease:
          oneOf:
            - { $ref: "#/components/schemas/JobLease" }
            - { type: "null" }
        circuitBreaker: { $ref: "#/components/schemas/CircuitBreaker" }
        nextRunAt:
          oneOf:
            - { $ref: "#/components/schemas/Instant" }
            - { type: "null" }
        lastRunId:
          oneOf:
            - { $ref: "#/components/schemas/Id" }
            - { type: "null" }
        speaks:
          type: boolean
          description: Whether this kind may produce a proactive message.
        createdAt: { $ref: "#/components/schemas/Instant" }
        updatedAt: { $ref: "#/components/schemas/Instant" }
    JobPage:
      type: object
      required: [items, nextCursor, hasMore]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Job" }
        nextCursor: { type: [string, "null"] }
        hasMore: { type: boolean }
    RunOutcome:
      type: string
      enum: [success, failure, skipped, suspended, abandoned]
    RunStep:
      type: object
      required: [id, index, sessionId, numTurns, costUsd, outcome, summary, startedAt, finishedAt]
      description: |
        Persisted after every turn and before the next one starts, because a
        turn is atomic from the outside: either it completed and we recorded
        the result, the session id, the turn count and the cost, or it did
        not. Resume is `--resume` against the stored session id.
      properties:
        id: { $ref: "#/components/schemas/Id" }
        index: { type: integer, minimum: 0 }
        sessionId: { type: [string, "null"] }
        numTurns: { type: integer, minimum: 0 }
        costUsd: { type: number, minimum: 0 }
        outcome: { $ref: "#/components/schemas/RunOutcome" }
        summary: { type: [string, "null"] }
        startedAt: { $ref: "#/components/schemas/Instant" }
        finishedAt:
          oneOf:
            - { $ref: "#/components/schemas/Instant" }
            - { type: "null" }
    Run:
      type: object
      required:
        [id, jobId, kind, triggerInstant, actualInstant, latenessMs, outcome, spoke, turns, costUsd, summary, error, attempts, startedAt, finishedAt, steps]
      description: |
        Every run records the gap between *scheduled* and *actual*. That gap
        is the whole point — a reminder that fired late is a nuisance, and
        one that pretended to be on time is a lie.
      properties:
        id: { $ref: "#/components/schemas/Id" }
        jobId: { $ref: "#/components/schemas/Id" }
        kind: { $ref: "#/components/schemas/JobKind" }
        triggerInstant: { $ref: "#/components/schemas/Instant" }
        actualInstant:
          oneOf:
            - { $ref: "#/components/schemas/Instant" }
            - { type: "null" }
        latenessMs: { type: integer }
        outcome: { $ref: "#/components/schemas/RunOutcome" }
        spoke: { type: boolean }
        turns: { type: integer, minimum: 0 }
        costUsd: { type: number, minimum: 0 }
        summary:
          type: [string, "null"]
          description: The model's own one line about what it did. Absent for zero-turn runs.
        error: { type: [string, "null"] }
        attempts: { type: integer, minimum: 0 }
        startedAt: { $ref: "#/components/schemas/Instant" }
        finishedAt:
          oneOf:
            - { $ref: "#/components/schemas/Instant" }
            - { type: "null" }
        steps:
          type: array
          items: { $ref: "#/components/schemas/RunStep" }
    RunPage:
      type: object
      required: [items, nextCursor, hasMore]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Run" }
        nextCursor: { type: [string, "null"] }
        hasMore: { type: boolean }

    # ------------------------------------------------------------ logs ---
...